+++ crates/slipbox-rpc/src/lib.rs
@@ -10,6 +10,7 @@ pub const METHOD_INDEX: &str = "slipbox/index";
 pub const METHOD_INDEX_FILE: &str = "slipbox/indexFile";
 pub const METHOD_INDEXED_FILES: &str = "slipbox/indexedFiles";
 pub const METHOD_SEARCH_FILES: &str = "slipbox/searchFiles";
+pub const METHOD_SEARCH_OCCURRENCES: &str = "slipbox/searchOccurrences";
 pub const METHOD_GRAPH_DOT: &str = "slipbox/graphDot";
 pub const METHOD_SEARCH_NODES: &str = "slipbox/searchNodes";
 pub const METHOD_RANDOM_NODE: &str = "slipbox/randomNode";